django/simple_auth/requirements.txt
django
django-rest-framework
django-df-api-drf==0.0.9
django-df-auth==1.0.0-alpha.4
django-cors-headers
